+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Changed the signature of `ProcessPayload`.
 - Only Build the Protobuf state once during serialization.
 - Capella blocks are execution.
+- Fixed panic when http request to subscribe to event stream fails.
 
 ### Deprecated